TLM: Adjust PV based on two prints
I printed one with PV of 125312 and one with 127066.368 (the original
corrected for the early 1.4% error).  The second print was 99.94% at
ping 1 but down to 98.71% at ping 7 (which is at 4.5m, the point where
the heuristic algorithm could be affecting something).  The first print
was at 100.85% at ping 7.

The correction is based on looking at the total cummulative error in
the "actual" pulses (corrected pulses) and then trying to hit the
mid point between the two sets of errors.