Add a priming area
This allows me to ditch printing skirts which are often not
correctly priming in multiple material prints anyway.

As part of this change, I improved the "next_move_full"
logic to only produce a full move for the next actual
move (instead of a retraction/un-retraction).  I think that this
will improve the behaviour when exiting the transition block.