gcode: Handle kisslicer layer changes more gracefully & batch transitions
If we have a layer that drops in height try to safely merge it
with adjacent runs to avoid the problem.

IF the problem is unavoidable, abort processing rather than producing
broken transitions.

As a nice side effect of this change, we also merge the transitions of
adjacent layers that are using the same tool.  A new printer parameter,
max_layer_height (default = 0.8*nozzle) caps how far we can merge.